requirements.txt
google-generativeai
python-dotenv
typer[all]
pypdf
python-docx
chromadb
nltk
reportlab